requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
redis==5.0.1
pydantic==2.5.3
python-dotenv==0.21.1
celery==5.3.6
flower==2.0.1
requests==2.31.0
openai==1.3.0
httpx==0.27.2
anthropic==0.34.0
botok==0.8.7